collector/requirements.txt
sqlalchemy
alembic
psycopg2-binary
requests
